// types/Datadog.ts
export type UserFlow =
  | "load-experiences"
  | "master-data"
  | "load-raw"
  | "create-raw"
  | "update-raw"
  | "delete-raw"
  | "publish-raw"
  | "load-curation"
  | "update-curation"
  | "publish-curation"
  | "create-translation"
  | "update-translation"
  | "publish-translation"
  | "load-translation"
  | "load-media"
  | "upload-media"
  | "delete-media"
  | "update-media"
  | "publish-media"
  | "load-option"
  | "create-option"
  | "update-option"
  | "update-agenda-slot-capacity"
  | "toggle-agenda-slot"
  | "load-activity-log"
  | "load-i18n"
  | "create-premade-item"
  | "create-premade-category"
  | "update-masterdata-translation"
  | "get-distribution-content"
  | "advanced-search-result-translation-missing"
  | "load-revision"
  | "load-raw-asterix-service-and-modalities"
  | "roles-no-such-role"
  | "load-allotments";

export type UserAction = "refresh-activity-log";